iso: Seperate config into different file
