+++ docs-partials/builder/utm/iso/Config-not-required.mdx
@@ -1,4 +1,4 @@
-<!-- Code generated from the comments of the Config struct in builder/utm/iso/builder.go; DO NOT EDIT MANUALLY -->
+<!-- Code generated from the comments of the Config struct in builder/utm/iso/config.go; DO NOT EDIT MANUALLY -->
 
 - `boot_steps` ([][]string) - This is an array of tuples of boot commands, to type when the virtual
   machine is booted. The first element of the tuple is the actual boot
@@ -80,4 +80,4 @@
   the file extension. By default this is packer-BUILDNAME, where
   "BUILDNAME" is the name of the build.
 
-<!-- End of code generated from the comments of the Config struct in builder/utm/iso/builder.go; -->
+<!-- End of code generated from the comments of the Config struct in builder/utm/iso/config.go; -->
